Refactor duplicated per-session saga lock tracking from synchronized storage and the testable version
